\name{firer-package}
\alias{firer-package}
\alias{firer}
\docType{package}
\title{
\packageTitle{firer}
}
\description{
\packageDescription{firer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{firer}
\packageIndices{firer}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{firer}

Maintainer: \packageMaintainer{firer}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
