\name{fireR-package}
\alias{fireR-package}
\alias{fireR}
\docType{package}
\title{
\packageTitle{fireR}
}
\description{
\packageDescription{fireR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fireR}
\packageIndices{fireR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fireR}

Maintainer: \packageMaintainer{fireR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
